requirements.txt
# Use PyTorch CPU wheels
--index-url https://download.pytorch.org/whl/cpu
# PyG prebuilt wheels matching torch==2.3.1+cpu
-f https://data.pyg.org/whl/torch-2.3.1+cpu.html

# Core
torch==2.3.1
torch-geometric==2.5.3
pyg-lib==0.4.0
torch-scatter==2.1.2
torch-sparse==0.6.18
torch-cluster==1.6.3
torch-spline-conv==1.2.2

# Scientific stack
numpy
pandas
scipy
matplotlib
seaborn

# Cheminformatics
rdkit-pypi
pubchempy